requirements.txt
qdrant-client
python-dotenv
langchain
langgraph
langchain-nomic
langchain-community
bs4
docx2txt